pyproject.toml
[tool.poetry]
name = "final_proj"
version = "0.1.0"
description = ""
authors = ["l4blee <l4blee@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.52.0"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
